+++ backend/src/application/use_cases/user_ops/delete_user.py
@@ -1,8 +1,7 @@
-from sqlalchemy.orm import Session
-
 from backend.src.application.interfaces.user_repository import UserRepository
-from backend.src.application.use_cases import user_ops
 
 
 def delete_user(user_repo: UserRepository, user_id: int) -> bool:
+    if user_repo.get_by_id(user_id) is None:
+        raise ValueError("User not found")
     return user_repo.delete(user_id=user_id)
